Add styled components
