+++ src/components/Photo/Photo.Styles.js
@@ -1,0 +1,5 @@
+import styled from 'styled-components';
+
+export const Image = styled.img`
+  width: 100%;
+`;
