+++ src/pages/Home/Home.Styles.js
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+export const Wrapper = styled.div`
+  width: 50%;
+  margin: 0 auto;
+`;
